Must start from title screen to test
Script compile order needs to be changed to make sure we don't get errors with the dialogue on loading the beach scene. Wetland scene needs to be changed to the new perspective.
Attempted to put UI elements all in their own group. Needs to be tested to make sure that didn't break anything.
Camera is not properly moving vertically.